requirements.txt
paho-mqtt
prometheus-client